GraphDrawer has been substantially rewritten so that it is now an ItemDisplayer just like the HTML View or Text View. It now has a displayItem method and should now be tabifiable (tm).
getGraphImage is gone; it is now perfectly ok for a GraphDrawer to
be reused and draw new graphs of new functions and whatnot.

drawGraphImage is gone. In its place, there are now two virtual methods,
drawModuleGraph and drawFunctionGraph. They return NULL if you don't override
them. If the particular GraphDrawer supports graphing a function, it should
override drawFunctionGraph; likewise for drawModuleGraph.