Pull r63760 into Dib:
Avoids generating a legalization assert for the case where a vector type is
legal but when legalizing the operation, we split the vector type and generate a
library call whose type needs to be promoted. For example, X86 with SSE on but
MMX off, a divide v2i64 will be scalarized to 2 calls to a library using i64.