--- Merging (from foreign repository) r67093 into '.': A    test/CodeGen/X86/x86-64-malloc.ll U    lib/CodeGen/SelectionDAG/SelectionDAGBuild.cpp U    test/CodeGen/X86/alloca-align-rounding.ll
Fix codegen to compute the size of an allocation by multiplying the
size by the array amount as an i32 value instead of promoting from
i32 to i64 then doing the multiply.  Not doing this broke wrap-around
assumptions that the optimizers (validly) made.  The ultimate real
fix for this is to introduce i64 version of alloca and remove mallocinst.

This fixes PR3829