Implement x86 h-register extract support.  - Add patterns for h-register extract, which avoids a shift and mask,    and in some cases a temporary register.  - Add address-mode matching for turning (X>>(8-n))&(255<<n), where    n is a valid address-mode scale value, into an h-register extract    and a scaled-offset address.  - Replace X86's MOV32to32_ and related instructions with the new    target-independent COPY_TO_SUBREG instruction.
On x86-64 there are complicated constraints on h registers, and
CodeGen doesn't currently provide a high-level way to express all of them,
so they are handled with a bunch of special code. This code currently only
supports extracts where the result is used by a zero-extend or a store,
though these are fairly common.

These transformations are not always beneficial; since there are only
4 h registers, they sometimes require extra move instructions, and
this sometimes increases register pressure because it can force out
values that would otherwise be in one of those registers. However,
this appears to be relatively uncommon.