--- Merging (from foreign repository) r69585 into '.': U    test/CodeGen/X86/2008-08-05-SpillerBug.ll A    test/CodeGen/X86/2009-04-20-LinearScanOpt.ll U    test/CodeGen/X86/2008-10-16-SpillerBug.ll U    lib/CodeGen/LiveIntervalAnalysis.cpp U    lib/CodeGen/RegAllocLinearScan.cpp
Added a linearscan register allocation optimization. When the register allocator
spill an interval with multiple uses in the same basic block, it creates a
different virtual register for each of the reloads. e.g.

      %reg1498<def> = MOV32rm %reg1024, 1, %reg0, 12, %reg0, Mem:LD(4,4) [sunkaddr39 + 0]
        %reg1506<def> = MOV32rm %reg1024, 1, %reg0, 8, %reg0, Mem:LD(4,4) [sunkaddr42 + 0]
        %reg1486<def> = MOV32rr %reg1506
        %reg1486<def> = XOR32rr %reg1486, %reg1498, %EFLAGS<imp-def,dead>
        %reg1510<def> = MOV32rm %reg1024, 1, %reg0, 4, %reg0, Mem:LD(4,4) [sunkaddr45 + 0]

=>

        %reg1498<def> = MOV32rm %reg2036, 1, %reg0, 12, %reg0, Mem:LD(4,4) [sunkaddr39 + 0]
        %reg1506<def> = MOV32rm %reg2037, 1, %reg0, 8, %reg0, Mem:LD(4,4) [sunkaddr42 + 0]
        %reg1486<def> = MOV32rr %reg1506
        %reg1486<def> = XOR32rr %reg1486, %reg1498, %EFLAGS<imp-def,dead>
        %reg1510<def> = MOV32rm %reg2038, 1, %reg0, 4, %reg0, Mem:LD(4,4) [sunkaddr45 + 0]

From linearscan's point of view, each of reg2036, 2037, and 2038 are separate
registers, each is "killed" after a single use. The reloaded register is
available and it's often clobbered right away. e.g. In thise case reg1498 is
allocated EAX while reg2036 is allocated RAX. This means we end up with multiple
reloads from the same stack slot in the same basic block.

Now linearscan recognize there are other reloads from same SS in the same BB. So
it'll "downgrade" RAX (and its aliases) after reg2036 is allocated until the
next reload (reg2037) is done. This greatly increase the likihood reloads from
SS are reused.

This speeds up sha1 from OpenSSL by 5.8%. It is also an across the board win for
SPEC2000 and 2006.