This commit fixes 254.gap on Mac OS X and should fix PR#8103.
First, when finding which arguments need pools, process an entire equivalence
class at a time.  This ensures that every function in an equivalence class has
the same DSNodes listed in MarkedNodes and ArgNodes in its FuncInfo, which
helps to transform the functions in the equivalence class uniformly.

Second, when transforming a call site, examine the target with the maximum
number of arguments.  This originaly helped prevent eliding a pool argument
because a particular target didn't need it.  I'm not convinced that this change
in TransformFunctionBody.cpp is needed (or correct), but it's not breaking
anything at the moment, so it goes in.