XCore target: Add target specific flags to root Makefiles.
The default value of XCORE_TARGET_NEEDS_MEMORY is 32 (MB).
It should be noted that the XCore emulator's speed is affected by the size of memory.
Also the amount of host memory required is several times XCore-image memory size.
Thus running several large xcore-images in parallel may result in thrashing!
The default of 32 will require ~300MB of host memory.

As Makefile.config.in is included multiple times (viz: include $(LEVEL)/Makefile.config),
the addition of XCORE_TARGET_FLAGS to X_TARGET_FLAGS is done in Makefile.test

RUNTIMELIMIT is set to a default of 1 hour for the XCore emulator.
Test that take around or more than 1 hour to run have their RUNTIMELIMIT extended.